+++ packages/identity-cli/src/commands/verificationMethodAdd.ts
@@ -1,13 +1,19 @@
 // Copyright 2024 IOTA Stiftung.
 // SPDX-License-Identifier: Apache-2.0.
-import { CLIDisplay, CLIOptions, CLIParam, CLIUtils, type CliOutputOptions } from "@gtsc/cli-core";
-import { Converter, I18n, Is, StringHelper } from "@gtsc/core";
-import { IotaIdentityConnector, IotaIdentityUtils } from "@gtsc/identity-connector-iota";
-import { DidVerificationMethodType } from "@gtsc/standards-w3c-did";
-import { VaultConnectorFactory } from "@gtsc/vault-models";
+import {
+	CLIDisplay,
+	CLIOptions,
+	CLIParam,
+	CLIUtils,
+	type CliOutputOptions
+} from "@twin.org/cli-core";
+import { Converter, I18n, Is, StringHelper } from "@twin.org/core";
+import { IotaIdentityConnector, IotaIdentityUtils } from "@twin.org/identity-connector-iota";
+import { DidVerificationMethodType } from "@twin.org/standards-w3c-did";
+import { VaultConnectorFactory } from "@twin.org/vault-models";
 
-import { IotaWalletConnector } from "@gtsc/wallet-connector-iota";
-import { WalletConnectorFactory } from "@gtsc/wallet-models";
+import { IotaWalletConnector } from "@twin.org/wallet-connector-iota";
+import { WalletConnectorFactory } from "@twin.org/wallet-models";
 import { Command, Option } from "commander";
 import { setupVault } from "./setupCommands";