+++ packages/identity-models/src/models/api/IIdentityProfileGetResponse.ts
@@ -1,6 +1,6 @@
 // Copyright 2024 IOTA Stiftung.
 // SPDX-License-Identifier: Apache-2.0.
-import type { IJsonLdDocument } from "@gtsc/data-json-ld";
+import type { IJsonLdDocument } from "@twin.org/data-json-ld";
 
 /**
  * Response to get an identity details.