+++ packages/identity-models/src/models/api/IIdentityResolveResponse.ts
@@ -1,6 +1,6 @@
 // Copyright 2024 IOTA Stiftung.
 // SPDX-License-Identifier: Apache-2.0.
-import type { IDidDocument } from "@gtsc/standards-w3c-did";
+import type { IDidDocument } from "@twin.org/standards-w3c-did";
 
 /**
  * Response to resolving an identity.