+++ packages/identity-connector-entity-storage/docs/changelog.md
@@ -1,5 +1,5 @@
-# @gtsc/identity-connector-entity-storage- Changelog
+# @twin.org/identity-connector-entity-storage- Changelog
 
-## v0.0.1
+## v0.0.1-next.3
 
 - Initial Release