+++ packages/identity-service/docs/changelog.md
@@ -1,5 +1,5 @@
-# @gtsc/identity-service - Changelog
+# @twin.org/identity-service - Changelog
 
-## v0.0.1
+## v0.0.1-next.3
 
 - Initial Release